Superhirn PATCH by Florian Zeller: Removed dead code, fixed implicit conversions
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@10084 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/TRANSFORMATIONS/FEATUREFINDER/SUPERHIRN/feature.C
@@ -601,7 +601,7 @@ void feature::add_MS2_info(map<double, vector<ms2_info> >* IN){
 //////////////////////////////////////////////////
 // check how many matches
 int feature::get_nb_common_match(){
-  return matched_feature_list.size() + 1;
+  return (int) matched_feature_list.size() + 1;
 }
 
 
@@ -652,7 +652,7 @@ void feature::add_matched_feature(feature* IN){
   int insert_ID = IN->get_spectrum_ID();
   map<int, feature>::iterator F = matched_feature_list.find( insert_ID );
   if( F != matched_feature_list.end() ){
-    insert_ID += matched_feature_list.size();
+    insert_ID += (int) matched_feature_list.size();
   }
   matched_feature_list.insert( pair<int,feature>( insert_ID, *IN ) );
   IN = NULL;
